start refactoring ...
